streaming_pipeline/kafka/requirements.txt
pyspark>=3.5.0
google-cloud-storage>=2.10.0
pandas>=2.0.0
pyarrow>=13.0.0
